Move generic CSV reading methods to a Mixin
This will DRY up the commands that all read from a CSV file in some way.

Might even be useful outside of this project.
